third_party/blink/web_tests/virtual/css-sizing-keyword-animation/external/wpt/css/css-sizing/animation/min-height-composition-expected.txt

This is a testharness.js-based test.
[FAIL] Compositing: property <min-height> underlying [100px] from add [100px] to add [auto] at (-0.3) should be [200px]
  assert_equals: expected "200px " but got "calc - size ( auto , 100px + 130px + size * - 0.3 ) "
[FAIL] Compositing: property <min-height> underlying [100px] from add [100px] to add [auto] at (0) should be [200px]
  assert_equals: expected "200px " but got "calc - size ( auto , 100px + 100px + size * 0 ) "
[FAIL] Compositing: property <min-height> underlying [100px] from add [100px] to add [auto] at (0.5) should be [auto]
  assert_equals: expected "0px " but got "calc - size ( auto , 100px + 50px + size * 0.5 ) "
[FAIL] Compositing: property <min-height> underlying [100px] from add [100px] to add [auto] at (1) should be [auto]
  assert_equals: expected "0px " but got "calc - size ( auto , 100px + 0px + size * 1 ) "
[FAIL] Compositing: property <min-height> underlying [100px] from add [100px] to add [auto] at (1.5) should be [auto]
  assert_equals: expected "0px " but got "calc - size ( auto , 100px - 50px + size * 1.5 ) "
Harness: the test ran to completion.
